Add infrastructure to automatically test integration
This commits add a test-case project and the Tycho
infrastructure enabling the launch of these test cases
on top of an arbitrary product.

This is intended to be used as a validation of the EPP
build result checking pain points of integrations.

This commit also introduce a tests which check that
every EPackage is resolvable and valid.