Prevent exception on each selection change.
When the view gains focus, getExplorerContextualViewID() gets called to
return an ID of view which should be opened.
Right now this returns "" and as such Eclipse then fails with an
exception. 
This commit make sure the getExplorerContextualViewID() method is not
static (otherwise I don't see the point as then nobody could override
it) and that if the returned string is null or empty then we won't ask
Eclipse to show the view.