[releng] Add JIRO-specific publication script
This is identical to the publish-nightly.sh version, except that it
uses SSH/SCP for all manipulations of files and folders on
download.eclipse.org.

Note that at the moment it publishes into a "staging" folder instead
of the normal "updates". This is to allow for testing without risking
conflicts with the legacy JIPP's jobs. It will need to be modified to
"updates" when the new JIPP becomes the official one.

Change-Id: I64b0b8d0feb6a2ac24734b65ae655f116b491166
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>