[cleanup] Use RuntimeLoggerManager in NewRepresentationAction
NewRepresentationAction.computeDefaultName was the only caller of
SiriusPlugin whose package has change for Sirius 7.0.0.

Use RuntimeLoggerManager to handle the EvaluationException thrown by the
evaluation of DiagramDescription::titleExpression. The code is now
consistent with the Sirius code base.

This allows to remove the 7.0.0 minimum version set on Sirius dependency
in the previous commit.

Bug: http://eclip.se/531282
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>